# Conflicts:
#	pubspec.lock

+++ lib/homeScreen.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
+import 'package:flutter/material.dart';
 import 'package:sqflite/sqflite.dart';
 import 'userData.dart';
 import 'hSimpleTable.dart';
@@ -26,25 +27,19 @@ class _HomeScreenState extends State<HomeScreen> {
     width: 40,
     height: 40,
   );
-  late List<bool> cmdStatus;
+  late List<bool> cmdStatus =
+      List.generate(cmdCaptions.length, (index) => true);
 
   // (cmdStatus[entry.key] = true) ? wrightIcon : wrongIcon
 
-  late List<ElevatedButton> btnList = cmdCaptions
-      .asMap()
-      .entries
-      .map(
-        (entry) => ElevatedButton(
-            onPressed: () => btnClicked(entry.key),
-            child: Text(cmdCaptions[entry.key])),
-      )
-      .toList();
+  late List<OutlinedButton> btnList;
 
   late List<Image> chkList;
 
   @override
   void initState() {
-    myUser.initData(100);
+    //myUser.initData(100);
+
     cmdCaptions = [
       'Create DB',
       'Create Table',
@@ -54,10 +49,13 @@ class _HomeScreenState extends State<HomeScreen> {
       'Delete Data',
       'Delete Table',
       'Delete DB',
-      'Clear Scr Data'
+      'Clear Scr Data',
+      'Generate New Data'
     ];
-    cmdStatus = List.generate(cmdCaptions.length, (index) => false);
 
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      btnClicked(4);
+    });
     super.initState();
   }
 
@@ -68,14 +66,39 @@ class _HomeScreenState extends State<HomeScreen> {
         .entries
         .map((e) => (cmdStatus[e.key] == true) ? wrightIcon : wrongIcon)
         .toList();
+    TextStyle TS1=TextStyle(color: Colors.black);
+    TextStyle TS2=TextStyle(color: Colors.black26);
+
+    btnList = cmdCaptions
+        .asMap()
+        .entries
+        .map(
+          (entry) => OutlinedButton(
+          onPressed: () => btnClicked(entry.key),
+          child: Text(
+            style:
+            cmdStatus[entry.key]? TS1:TS2
+
+            ,cmdCaptions[entry.key],
+          )
+    ),
+    ).toList();
 
     return Scaffold(
       appBar: AppBar(
         title: Text(widget.title),
       ),
+      onDrawerChanged: (changed) {
+        if (changed == true) {
+          print('do all checks');
+          doAllChecks();
+        }
+
+        print(changed.toString());
+      },
       drawer: SafeArea(
         child: Container(
-          width: 200,
+          width: 250,
           decoration: BoxDecoration(
             border: Border.all(
               width: 10,
@@ -84,7 +107,7 @@ class _HomeScreenState extends State<HomeScreen> {
             borderRadius: const BorderRadius.all(
               Radius.circular(10),
             ),
-            color: Colors.cyanAccent,
+            color: Colors.white,
           ),
           child: Row(
             children: [
@@ -92,7 +115,7 @@ class _HomeScreenState extends State<HomeScreen> {
                 mainAxisAlignment: MainAxisAlignment.spaceAround,
                 children: btnList,
               ),
-              SizedBox(
+              const SizedBox(
                 width: 10,
               ),
               Column(
@@ -112,14 +135,75 @@ class _HomeScreenState extends State<HomeScreen> {
   }
 
   //////Methods
+  void doAllChecks() async {
+    var databasesPath = await getDatabasesPath();
+    var dbFilePath = '$databasesPath/my_dbase.db';
+    var dbExists = File(dbFilePath).existsSync();
+    if (dbExists == false) {
+      print('no such database');
+      cmdStatus[0] = true;
+      for (int i = 1; i < 8; i++) {
+        cmdStatus[i] = false;
+      }
+      return;
+    }
+    else{cmdStatus[0]=false;
+        }
+
+    setState(() {
+      cmdStatus;
+    });
+
+    late Database db;
+    db = await openDatabase('my_dbase.db');
+    if (db.isOpen == false) {
+      print('cant open database');
+      return;
+    }
+    var tables = await db
+        .rawQuery('SELECT * FROM sqlite_master WHERE name="datatable";');
+
+    if (tables.isEmpty) {
+      // Create the table
+      print('no such table');
+      setState(() {
+        cmdStatus[1]=true;
+      });
+      return;
+    }else{setState(() {
+  cmdStatus[1]=false;
+});
+}
+
+    List<Map>? gotlist = await db.database.rawQuery('SELECT * FROM datatable');
+
+    if (myUser.userInfo.length == 0) {
+      print('Empty Screen');
+      setState(() {
+        cmdStatus[2]=false;
+      });
+      return;
+    }else{
+      setState(() {
+        cmdStatus[2]=true;
+
+      });
+      return;
+
+    }
+  }
+
   void btnClicked(int idx) async {
     var databasesPath = await getDatabasesPath();
     var dbFilePath = '$databasesPath/my_dbase.db';
-    var dbExists = await File(dbFilePath).existsSync();
+    var dbExists = File(dbFilePath).existsSync();
     late Database db;
 
     switch (idx) {
       case 0:
+        if (cmdStatus[0] == false) {
+          return;
+        }
         if (dbExists) {
           setState(() {
             cmdStatus[0] = true;
@@ -139,7 +223,11 @@ class _HomeScreenState extends State<HomeScreen> {
           }
         }
         break; //Create Database
+
       case 1:
+        if (cmdStatus[1] == false) {
+          return;
+        }
         if (dbExists) {
           db = await openDatabase('my_dbase.db');
           try {
@@ -165,7 +253,11 @@ class _HomeScreenState extends State<HomeScreen> {
           cmdStatus[1] = true;
         });
         break;
+
       case 2:
+        if (cmdStatus[2] == false) {
+          return;
+        }
         print('started inserting data');
         db = await openDatabase('my_dbase.db');
 
@@ -183,23 +275,38 @@ class _HomeScreenState extends State<HomeScreen> {
 
         break;
 
-
       // reading data
       case 4:
+        if (cmdStatus[4] == false) {
+          return;
+        }
         db = await openDatabase('my_dbase.db');
-        List<Map> gotlist =
+        var tables = await db
+            .rawQuery('SELECT * FROM sqlite_master WHERE name="datatable";');
+
+        if (tables.isEmpty) {
+          // Create the table
+          print('no such table');
+          break;
+        }
+
+        List<Map>? gotlist =
             await db.database.rawQuery('SELECT * FROM datatable');
+        print(gotlist.length.toString());
+        if (gotlist.length == 0) {
+          print('Empty Table');
+          break;
+        }
         print(gotlist.length);
         //print(gotlist);
         print(gotlist[0]);
         //print(myUser.userInfo[0]);
         myUser.userInfo.clear();
-        if(gotlist.length>0){
+        if (gotlist.length > 0) {
           setState(() {
-           myUser.insertData(gotlist);
+            myUser.insertData(gotlist);
           });
         }
-
 
         print(myUser.userInfo.length);
         print('added all');
@@ -208,7 +315,29 @@ class _HomeScreenState extends State<HomeScreen> {
         });
         break;
 
+      /// delete data from db
+      case 5:
+        if (cmdStatus[5] == false) {
+          return;
+        }
+        db = await openDatabase('my_dbase.db');
+        await db.database.rawQuery('DELETE FROM datatable');
+        print('deleted all rows');
+
+        break;
+      case 6:
+        if (cmdStatus[6] == false) {
+          return;
+        }
+        db = await openDatabase('my_dbase.db');
+        await db.database.rawQuery('DROP TABLE IF EXISTS datatable');
+        print('deleted the table');
+        break;
+
       case 7:
+        if (cmdStatus[7] == false) {
+          return;
+        }
         await databaseFactory.deleteDatabase(dbFilePath);
         dbExists = await File(dbFilePath).existsSync();
         if (!dbExists) {
@@ -222,11 +351,22 @@ class _HomeScreenState extends State<HomeScreen> {
 
         break;
       case 8:
+        if (cmdStatus[8] == false) {
+          return;
+        }
         //delete table screen
         cmdStatus[8] = true;
         setState(() {
           myUser.userInfo.clear();
         });
+
+        break;
+      case 9:
+        if (cmdStatus[9] == false) {
+          return;
+        }
+        myUser.initData(100);
+        setState(() {});
 
         break;
